+++ src/pages/cadastro/CadastroPage.styles.ts
@@ -64,6 +64,14 @@ export const I = styled.img `
     box-sizing: border-box;
 `
 
+export const Eyes = styled.img`
+  position: absolute;
+  height: 1.44rem;
+  top: 0.38rem;
+  right: 0.48rem;
+  box-sizing: border-box;
+`;
+
 export const Checkbox = styled.p `
     display: flex;
     align-items: center;