+++ src/pages/fasesDeTratamento/FasesDeTratamentoPage.styles.ts
@@ -40,27 +40,6 @@ export const ConteinerCheckBox = styled.div`
   gap: 0.563rem;
 `;
 
-export const ContainerCheckbox = styled.div`
-  .checkbox {
-    display: none;
-  }
-  .checkbox + label:before {
-    content: "";
-    display: inline-block;
-    width: 1rem;
-    height: 1rem;
-    border-radius: 0.3rem;
-    margin-right: 0.5rem;
-    border: 2px solid ${(props) => props.theme.colors.variant3};
-    background: ${(props) => props.theme.colors.buttonOff};
-  }
-
-  .custom-checkbox input:checked + label:before {
-    background-image: url("../../../src/assets/Vector.png");
-    background-position: center;
-  }
-`;
-
 export const ButtonLocal = styled.div`
   margin-top: 40px;
 `;